feat(ui): migrate VirtualList to src/components/ui/
