+++ src/components/common/VirtualList.tsx
@@ -1,267 +1,3 @@
-/**
- * VirtualList — high-performance virtualized scrolling list for SolidJS.
- *
- * Only renders the rows visible in the viewport plus an overscan buffer.
- * Supports a fixed row height (required for O(1) index calculation) and
- * provides an optional auto-scroll-to-bottom behaviour.
- *
- * Performance characteristics:
- *  - DOM nodes:  O(visible rows + 2×OVERSCAN)  instead of O(total)
- *  - Scroll FPS: 60 even with 100K entries
- *  - Layout reads: zero per frame (no scrollHeight queries in hot path)
- *
- * Usage:
- *   <VirtualList
- *     items={filteredEntries()}
- *     rowHeight={20}
- *     renderRow={(item, index) => <MyRow entry={item} />}
- *     autoScroll={autoScroll()}
- *     onScrolledToBottom={() => setAutoScroll(true)}
- *     onScrolledUp={() => setAutoScroll(false)}
- *   />
- */
-
-import {
-  type JSX,
-  For,
-  createEffect,
-  createMemo,
-  createSignal,
-  onCleanup,
-  onMount,
-} from "solid-js";
-
-export interface VirtualListHandle {
-  scrollToBottom: () => void;
-}
-
-export interface VirtualListProps<T> {
-  /** The full (filtered) array to display. */
-  items: T[];
-  /** Height of every row in pixels. Must be fixed. */
-  rowHeight: number;
-  /** Render a single row. Receives the item and its absolute index in `items`. */
-  renderRow: (item: T, index: number) => JSX.Element;
-  /**
-   * When true the list scrolls to the bottom whenever items change and the
-   * user is already at the bottom (or auto-scroll has never been overridden).
-   */
-  autoScroll?: boolean;
-  /** Called when the user scrolls to within `bottomThreshold` px of the end. */
-  onScrolledToBottom?: () => void;
-  /** Called when the user scrolls up away from the bottom. */
-  onScrolledUp?: () => void;
-  /** How many extra rows to render above and below the viewport. Default 30. */
-  overscan?: number;
-  /**
-   * Distance from the bottom (in px) that counts as "at the bottom".
-   * Default 40 (two rows).
-   */
-  bottomThreshold?: number;
-  /** CSS class applied to the outer scrolling container. */
-  class?: string;
-  /** Inline styles applied to the outer scrolling container. */
-  style?: JSX.CSSProperties;
-  /**
-   * When set to a non-null number, the list scrolls to that absolute item
-   * index.  Wrap in a signal: `jumpTo={jumpTarget()}`; update the signal to
-   * trigger a scroll.  The list sets `autoScroll` to false on jump.
-   */
-  jumpTo?: number | null;
-  /**
-   * Imperative handle callback. The parent receives a `VirtualListHandle`
-   * object on mount, enabling programmatic `scrollToBottom()` calls.
-   */
-  handle?: (api: VirtualListHandle) => void;
-  /**
-   * Cumulative pixel offset used to compensate for front-eviction scroll
-   * jumps. Each time entries are removed from the front of the list the
-   * parent increments this value by `removedCount × rowHeight`; the list
-   * applies only the delta to `scrollTop` so the viewport stays stable.
-   */
-  scrollCompensate?: number;
-}
-
-export function VirtualList<T>(props: VirtualListProps<T>): JSX.Element {
-  const OVERSCAN = () => props.overscan ?? 30;
-  const THRESHOLD = () => props.bottomThreshold ?? 40;
-
-  // ── Container measurement ─────────────────────────────────────────────────
-
-  let outerRef!: HTMLDivElement;
-  const [containerHeight, setContainerHeight] = createSignal(600);
-  const [scrollTop, setScrollTop] = createSignal(0);
-
-  onMount(() => {
-    // Measure once synchronously.
-    setContainerHeight(outerRef.clientHeight || 600);
-
-    // Keep height in sync when the panel is resized.
-    const ro = new ResizeObserver((entries) => {
-      const h = entries[0]?.contentRect.height;
-      if (h && h > 0) setContainerHeight(h);
-    });
-    ro.observe(outerRef);
-    onCleanup(() => ro.disconnect());
-
-    // Expose imperative scroll-to-bottom API to the parent.
-    props.handle?.({
-      scrollToBottom: () => {
-        outerRef.scrollTop = outerRef.scrollHeight;
-      },
-    });
-  });
-
-  // ── Windowing memos ───────────────────────────────────────────────────────
-
-  const totalCount = () => props.items.length;
-  const totalHeight = () => totalCount() * props.rowHeight;
-
-  const startIndex = createMemo(() =>
-    Math.max(0, Math.floor(scrollTop() / props.rowHeight) - OVERSCAN())
-  );
-
-  const endIndex = createMemo(() =>
-    Math.min(
-      totalCount(),
-      Math.ceil((scrollTop() + containerHeight()) / props.rowHeight) + OVERSCAN()
-    )
-  );
-
-  const visibleItems = createMemo(() =>
-    props.items.slice(startIndex(), Math.max(startIndex(), endIndex()))
-  );
-
-  // Pixel offset of the first rendered row from the top of the inner div.
-  const offsetY = () => startIndex() * props.rowHeight;
-
-  // ── Auto-scroll ───────────────────────────────────────────────────────────
-
-  // Track whether we were at the bottom before each batch so we can decide
-  // whether to scroll after the DOM updates.
-  let wasAtBottom = true;
-
-  // Called whenever items change AND autoScroll is requested.
-  // We schedule via queueMicrotask so the DOM has updated before we read
-  // scrollHeight (once, not on every scroll event).
-  const scheduleAutoScroll = () => {
-    if (props.autoScroll && wasAtBottom) {
-      queueMicrotask(() => {
-        if (outerRef) {
-          outerRef.scrollTop = outerRef.scrollHeight;
-        }
-      });
-    }
-  };
-
-  // Watch items length changes for auto-scroll.
-  createEffect(() => {
-    const count = props.items.length;
-    if (count > 0) scheduleAutoScroll();
-  });
-
-  // ── jumpTo effect ─────────────────────────────────────────────────────────
-
-  createEffect(() => {
-    const idx = props.jumpTo;
-    if (idx === null || idx === undefined || !outerRef) return;
-    // Center the target row in the viewport
-    const targetTop = idx * props.rowHeight;
-    const viewportMid = containerHeight() / 2;
-    outerRef.scrollTop = Math.max(0, targetTop - viewportMid + props.rowHeight / 2);
-    wasAtBottom = false;
-  });
-
-  // ── Front-eviction scroll compensation ───────────────────────────────────
-
-  // Each instance keeps its own counter so compensation is not shared across
-  // multiple VirtualList renders in the same page.
-  let prevCompensate = 0;
-  createEffect(() => {
-    const current = props.scrollCompensate ?? 0;
-    const delta = current - prevCompensate;
-    prevCompensate = current;
-    if (delta !== 0 && outerRef) {
-      outerRef.scrollTop += delta;
-    }
-  });
-
-  // ── Scroll handler ────────────────────────────────────────────────────────
-
-  function handleScroll() {
-    if (!outerRef) return;
-    const st = outerRef.scrollTop;
-    setScrollTop(st);
-
-    const distFromBottom =
-      outerRef.scrollHeight - st - outerRef.clientHeight;
-    const atBottom = distFromBottom < THRESHOLD();
-
-    if (atBottom && !wasAtBottom) {
-      wasAtBottom = true;
-      props.onScrolledToBottom?.();
-    } else if (!atBottom && wasAtBottom) {
-      wasAtBottom = false;
-      props.onScrolledUp?.();
-    }
-  }
-
-  // ── Render ────────────────────────────────────────────────────────────────
-
-  return (
-    <div
-      ref={outerRef}
-      onScroll={handleScroll}
-      class={props.class}
-      style={{
-        "overflow-y": "auto",
-        "overflow-x": "hidden",
-        position: "relative",
-        ...(props.style ?? {}),
-      }}
-    >
-      {/* Full-height spacer so the scroll bar reflects total content height */}
-      <div
-        style={{
-          height: `${totalHeight()}px`,
-          position: "relative",
-          "pointer-events": "none",
-        }}
-      />
-
-      {/* Rendered window — positioned absolutely over the spacer */}
-      <div
-        style={{
-          position: "absolute",
-          top: "0",
-          left: "0",
-          right: "0",
-          transform: `translateY(${offsetY()}px)`,
-          "pointer-events": "auto",
-          "will-change": "transform",
-        }}
-      >
-        {/*
-         * <For> reconciles by item identity — when new entries arrive at the
-         * bottom, only those new rows are created.  The existing ~60 visible
-         * rows are reused with zero DOM mutations, giving 60fps at any buffer
-         * size.  (.map() would destroy and recreate all 60 rows every batch.)
-         *
-         * Absolute-index stability proof: for an item that stays in the
-         * visible window, its absoluteIndex = startIndex + localIndex is
-         * invariant across a scroll of δ rows because startIndex increases by
-         * δ while localIndex decreases by δ.  So the captured value from the
-         * initial render is always correct.
-         */}
-        <For each={visibleItems()}>
-          {(item, getLocalIndex) =>
-            // eslint-disable-next-line solid/reactivity
-            props.renderRow(item, startIndex() + getLocalIndex())
-          }
-        </For>
-      </div>
-    </div>
-  );
-}
-
-export default VirtualList;
+// Re-export bridge — all new code should import from "@/components/ui"
+// This bridge will be removed in the App Integration plan (Phase 3)
+export { VirtualList, type VirtualListProps, type VirtualListHandle } from "@/components/ui/VirtualList";
