+++ src/components/ui/VirtualList/index.ts
@@ -1,0 +1,2 @@
+export { VirtualList } from "./VirtualList";
+export type { VirtualListProps, VirtualListHandle } from "./VirtualList";
