+++ src/components/ui/index.ts
@@ -1,3 +1,4 @@
 // Design system barrel
 export { Icon, type IconProps } from "./Icon";
 export { Resizable, type ResizableProps } from "./Resizable";
+export { VirtualList, type VirtualListProps, type VirtualListHandle } from "./VirtualList";
